refractor code
