+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/code/Code.java
@@ -1,19 +1,20 @@
-package com.chilleric.franchise_sys.repository.systemRepository.accessability;
+package com.chilleric.franchise_sys.repository.system_repository.code;
 
+import java.util.Date;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.core.mapping.Document;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Document(collection = "accessability")
+@Document(collection = "codes")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Accessability {
+public class Code {
   private ObjectId _id;
   private ObjectId userId;
-  private ObjectId targetId;
-  private boolean editable;
-  private boolean isServer;
+  private TypeCode type;
+  private String code;
+  private Date expiredDate;
 }
