+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/code/CodeRepository.java
@@ -1,4 +1,4 @@
-package com.chilleric.franchise_sys.repository.systemRepository.code;
+package com.chilleric.franchise_sys.repository.system_repository.code;
 
 import java.util.Optional;
 import org.bson.types.ObjectId;
