+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/language/LanguageRepository.java
@@ -1,8 +1,8 @@
-package com.chilleric.franchise_sys.repository.systemRepository.navbar;
+package com.chilleric.franchise_sys.repository.system_repository.language;
 
 import org.springframework.stereotype.Repository;
 import com.chilleric.franchise_sys.repository.abstract_repository.SystemRepository;
 
 @Repository
-public class NavbarRepository extends SystemRepository<Navbar> {
+public class LanguageRepository extends SystemRepository<Language> {
 }
