+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/navbar/NavbarRepository.java
@@ -1,8 +1,8 @@
-package com.chilleric.franchise_sys.repository.systemRepository.user;
+package com.chilleric.franchise_sys.repository.system_repository.navbar;
 
 import org.springframework.stereotype.Repository;
 import com.chilleric.franchise_sys.repository.abstract_repository.SystemRepository;
 
 @Repository
-public class UserRepository extends SystemRepository<User> {
+public class NavbarRepository extends SystemRepository<Navbar> {
 }
