+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/path/PathRepository.java
@@ -1,9 +1,8 @@
-package com.chilleric.franchise_sys.repository.systemRepository.settings;
+package com.chilleric.franchise_sys.repository.system_repository.path;
 
 import org.springframework.stereotype.Repository;
 import com.chilleric.franchise_sys.repository.abstract_repository.SystemRepository;
 
 @Repository
-public class SettingRepository extends SystemRepository<Setting> {
-
+public class PathRepository extends SystemRepository<Path> {
 }
