+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/permission/PermissionRepository.java
@@ -1,4 +1,4 @@
-package com.chilleric.franchise_sys.repository.systemRepository.permission;
+package com.chilleric.franchise_sys.repository.system_repository.permission;
 
 import static java.util.Map.entry;
 import java.lang.reflect.Field;
