+++ src/main/java/com/chilleric/franchise_sys/repository/system_repository/settings/SettingRepository.java
@@ -1,8 +1,9 @@
-package com.chilleric.franchise_sys.repository.systemRepository.path;
+package com.chilleric.franchise_sys.repository.system_repository.settings;
 
 import org.springframework.stereotype.Repository;
 import com.chilleric.franchise_sys.repository.abstract_repository.SystemRepository;
 
 @Repository
-public class PathRepository extends SystemRepository<Path> {
+public class SettingRepository extends SystemRepository<Setting> {
+
 }
