+++ src/main/java/com/chilleric/franchise_sys/service/navbar/NavbarServiceImpl.java
@@ -17,12 +17,12 @@ import com.chilleric.franchise_sys.dto.navbar.NavbarResponse;
 import com.chilleric.franchise_sys.dto.path.PathResponse;
 import com.chilleric.franchise_sys.exception.InvalidRequestException;
 import com.chilleric.franchise_sys.exception.ResourceNotFoundException;
-import com.chilleric.franchise_sys.repository.systemRepository.accessability.Accessability;
-import com.chilleric.franchise_sys.repository.systemRepository.navbar.ContentNavbar;
-import com.chilleric.franchise_sys.repository.systemRepository.navbar.Navbar;
-import com.chilleric.franchise_sys.repository.systemRepository.navbar.NavbarRepository;
-import com.chilleric.franchise_sys.repository.systemRepository.path.Path;
-import com.chilleric.franchise_sys.repository.systemRepository.path.PathRepository;
+import com.chilleric.franchise_sys.repository.system_repository.accessability.Accessability;
+import com.chilleric.franchise_sys.repository.system_repository.navbar.ContentNavbar;
+import com.chilleric.franchise_sys.repository.system_repository.navbar.Navbar;
+import com.chilleric.franchise_sys.repository.system_repository.navbar.NavbarRepository;
+import com.chilleric.franchise_sys.repository.system_repository.path.Path;
+import com.chilleric.franchise_sys.repository.system_repository.path.PathRepository;
 import com.chilleric.franchise_sys.service.AbstractService;
 
 @Service
