+++ src/main/java/com/chilleric/franchise_sys/service/permission/PermissionServiceImpl.java
@@ -24,10 +24,10 @@ import com.chilleric.franchise_sys.exception.InvalidRequestException;
 import com.chilleric.franchise_sys.exception.ResourceNotFoundException;
 import com.chilleric.franchise_sys.exception.UnauthorizedException;
 import com.chilleric.franchise_sys.repository.common_entity.ViewPoint;
-import com.chilleric.franchise_sys.repository.systemRepository.accessability.Accessability;
-import com.chilleric.franchise_sys.repository.systemRepository.permission.Permission;
-import com.chilleric.franchise_sys.repository.systemRepository.permission.PermissionRepository;
-import com.chilleric.franchise_sys.repository.systemRepository.user.UserRepository;
+import com.chilleric.franchise_sys.repository.system_repository.accessability.Accessability;
+import com.chilleric.franchise_sys.repository.system_repository.permission.Permission;
+import com.chilleric.franchise_sys.repository.system_repository.permission.PermissionRepository;
+import com.chilleric.franchise_sys.repository.system_repository.user.UserRepository;
 import com.chilleric.franchise_sys.service.AbstractService;
 import com.chilleric.franchise_sys.utils.DateFormat;
 import com.chilleric.franchise_sys.utils.ObjectUtilities;
