+++ src/main/java/com/chilleric/franchise_sys/service/settings/SettingServiceImpl.java
@@ -18,11 +18,11 @@ import com.chilleric.franchise_sys.dto.settings.ChangePasswordRequest;
 import com.chilleric.franchise_sys.dto.settings.SettingsRequest;
 import com.chilleric.franchise_sys.dto.settings.SettingsResponse;
 import com.chilleric.franchise_sys.exception.InvalidRequestException;
-import com.chilleric.franchise_sys.repository.systemRepository.language.LanguageRepository;
-import com.chilleric.franchise_sys.repository.systemRepository.settings.Setting;
-import com.chilleric.franchise_sys.repository.systemRepository.settings.SettingRepository;
-import com.chilleric.franchise_sys.repository.systemRepository.user.User;
-import com.chilleric.franchise_sys.repository.systemRepository.user.UserRepository;
+import com.chilleric.franchise_sys.repository.system_repository.language.LanguageRepository;
+import com.chilleric.franchise_sys.repository.system_repository.settings.Setting;
+import com.chilleric.franchise_sys.repository.system_repository.settings.SettingRepository;
+import com.chilleric.franchise_sys.repository.system_repository.user.User;
+import com.chilleric.franchise_sys.repository.system_repository.user.UserRepository;
 import com.chilleric.franchise_sys.service.AbstractService;
 import com.chilleric.franchise_sys.utils.DateFormat;
 import com.chilleric.franchise_sys.utils.PasswordValidator;
