+++ src/main/java/com/chilleric/franchise_sys/service/user/UserServiceImpl.java
@@ -24,10 +24,10 @@ import com.chilleric.franchise_sys.exception.InvalidRequestException;
 import com.chilleric.franchise_sys.exception.ResourceNotFoundException;
 import com.chilleric.franchise_sys.pusher.PusherResponse;
 import com.chilleric.franchise_sys.repository.common_entity.ViewPoint;
-import com.chilleric.franchise_sys.repository.systemRepository.accessability.Accessability;
-import com.chilleric.franchise_sys.repository.systemRepository.user.User;
-import com.chilleric.franchise_sys.repository.systemRepository.user.User.TypeAccount;
-import com.chilleric.franchise_sys.repository.systemRepository.user.UserRepository;
+import com.chilleric.franchise_sys.repository.system_repository.accessability.Accessability;
+import com.chilleric.franchise_sys.repository.system_repository.user.User;
+import com.chilleric.franchise_sys.repository.system_repository.user.User.TypeAccount;
+import com.chilleric.franchise_sys.repository.system_repository.user.UserRepository;
 import com.chilleric.franchise_sys.service.AbstractService;
 import com.chilleric.franchise_sys.utils.DateFormat;
